problem solving

+++ Array_Description.cpp
@@ -19,11 +19,13 @@ int main()
 {
     
     ios_base::sync_with_stdio(false);
-    int t; cin>>t;
-    while(t--){
+    
+    ll t; cin>>t;
+    while (t--)
+    {
         
     }
+    
 
     return 0;
 }
-
